backend/services/analytics/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
sqlalchemy>=2.0.35
alembic>=1.13.3
asyncpg>=0.29.0
python-dotenv>=1.0.1
pydantic>=2.9.2

-e ../../shared
psycopg2-binary>=2.9.11
